Rename Habits model to Habit to keep with convention
